Rework FnCallFileStatDetails().
Use the return from readlink() so as to set one '\0' byte instead of
filling a page with them so as to be sure the the first one readlink()
doesn't touch is '\0'.

Use strlcpy() rather than snprintf(,,"%s",).
Turn an else { if ... } into an else if, dedent lots of code.
Simplify some logic.  Split some long lines.  Reindented.