Structural simplifications in evalfunction.c
Make variables more local to avoid assignment in conditionals.
Code following an if (...) { ... return ...; } needn't be in an else
clause; may as well make it easy to see the function ends in a return.
Turn an else { if ... } into a simple else if.

(cherry picked from commit a28c5180a9f935cc600571e500b97202443925a8)