Remove the uninitialised buffer PRNG seed.
Use instead some rand() filled buffer. It is a bug if even that is not
enough, so complain loudly with UnexpectedError() (but try to continue).

In any case these kinds of hacks shouldn't be used, so print a
LOG_LEVEL_INFO message if that is necessary (e.g. on chroots that don't
have /dev/urandom). Still can be avoided by using WORKDIR/randseed.