Petty clean-up in conversion.c; and make IsRealNumber() sane.
Instead of initializing a value to a particular real and checking that
sscanf() changes it (which it won't when the given string happens to
be the given real), actually check the return from sscanf to know
whether it read a real.

Use the right cast on isspace()'s argument and simplify IsSpace().
Remove some wanton backslashes and put spaces where they belong.