Lots of petty clean-up in evalfunction.c
Make variables more local, split long lines, kill some pointless
comments.  After an if stanza that returns, don't bother putting
subsequent code into an else clause unless it helps readability.
Eliminate some pointless copying of values via intermediate buffers;
and unused argcount variable.  Skip some pointless initializers (of
variables about to be assigned to in the conditions of while loops).
Diverse simplifications of the logic of functions and avoidance of
duplications of clean-up.  Fold separate conditions with the same code
together with || to make visible the similarity of the checks.  Use
strlcpy() in preference to snprintf() when it suffices; at least use
the right buffer size when it doesn't.  Indent comments with code.
Skip lots of pointless setting of variables to what an initializer had
already set.

Reworked some looping over lines of files to condition the loop on
running out of lines, moving the EOF check out of the loop in the
process and avoiding duplication of clean-up code.  Fixed a leak (on
file read error) in one of those cases.