UMR-fix: EvalContext's ->pass was unset by default.
There's logging code (PromiseBanner) that reports the pass and
valgrind reported it as an uninitialized memory read.  So change
EvalContextNew() to use calloc and ditch a bunch of initializations
to NULL and false that this does for us.

(cherry picked from commit f8aa78dfcfe1ab327a94907b82057049cfe78a8c)

Conflicts:
	libpromises/eval_context.c