Warning-fix: eliminate uses of deprecated sprintf().
All in verify_databases.c; and clean up CreateTableColumns(), which
was one of the culprits.  It was also adding random extra amounts of
space to an allocation - "just in case" I suppose - thereby
obfuscating the actual amount of space needed.  Eliminated an
intermediate fixed-size buffer we didn't need (and any associated
truncation).  Avoid quadratic costs associated with repeatedly
traversing a string to find its end, at which to append.