Support or-entries not evaluated to a scalar
If any entry in an or-list was not evaluated to a scalar, consider it a
"class not set" and evaluate the rest of the entries.

Behavior change for classes promises that use an or constraint where
one of the list entries cannot be evaluated.

Implements redmine 6569
(cherry picked from commit bcab880c7fdb4e245a309a0129027d71c27ec1de)