Move time(NULL) call up the call-stack.
This is one step towards resolving the TODO that accompanies it: we
should really be using the time when PROCESSTABLE was populated.

This makes it possible for a unit test to not depend on the actual
time at which it's run, since it can now pass in the nominal time at
which to believe it's being run.