Assume promiser is a directory if depth_search is set
depth_search only makes sense if the promiser should be a directory
that might be copied from a remote server. So make sure we execute
the same code as if the policy writer expliclty asked for the promiser
to be created as a directory by suffixing with "/.".

Attempt to fix various redmine tickets reporting that copy_from
promises for remote directories create local files. Assumption is
that network interruption aborts transmission of data that tells
client that remote sources are a directory tree, and lacking the
'/.' in the promiser the client defaults to file creation.

Redmine #3400, 6027