Improved Log()ging and fixed a leak in evalfunction's JSON wrappers.
Successful parse yielding a non-container was conflated with failed
parsing; separate the two.  In the process, fix the leak of the json
object when it's not a container; and report *what* failed when we
didn't even get that.  Use JsonParseFile() in the one place that we're
just duplicating it's workings.