Test case for Redmine 6589.
Trying to delete a region where a specific select_end doesn't match
should be noop.

This is different from insert_lines, which appends the line at the
end if there is no match for select_end (see existing test case
10_files/09_insert_lines/302.cf). While it would be best to make
this behavior consistent, I'm opting for not rocking the boat and
keep existing behavior of insert_lines in place.