Redmine #6766: Work around bug in system headers on HPUX 11.23 IA64.
The definition using mpinfou inside /usr/include/machine/sys/getppdp.h
is supposed to be protected by an "#ifdef KERNEL", but it's missing.
Since fixing it properly requires changing the system, we instead fix
it by defining a dummy structure. Since it is a kernel structure it
will never actually be used anywhere.