Redmine #6602: Fix LMDB max_readers setting being too low under load.
We did not previously take cf-hub into account when calculating how
large the max_readers setting can be. In the worst case, every thread
in both cf-serverd and child in cf-hub can have it open at the same
time, as well as occasional cf-agents, so account for that.

Also, improve the error checking when setting the maximum readers.
LMDB will only succeed at setting the maxreaders limit if no other
database handles are open, including in other processes, despite
always returning success. This means that we should check afterwards
whether the value we tried to set was honored, and warn if it wasn't.
If there is a window of opportunity where the database is not open,
the problem will correct itself, but if the load is so high that the
database is constantly open, then the administrator may have to take
action and restart CFEngine.

Also moved this to the database opening code itself, so that it is
active for every database. This way we won't hit the problem again in
some future unforeseen database.
(cherry picked from commit 397aa1b52d626bbd23472e5517346e28e2709b6a)