Fix caching bug when both updating password and unlocking user.
The problem is that if we use /etc/passwd to get password hashes, we
are using fgetpwent(). This function uses a static buffer, which means
means we cannot call it again while we are still using its data, which
we are, all over the place in the code. However, if you update the
user hash, the value stored in that buffer becomes obsolete.

We cannot store a new value there, because we do not control the
buffer. We also cannot reread it, for the reasons mentioned above.

Instead, we take advantage of the fact that if we first set the
password, then lock, it doesn't matter what the password is, since it
will be locked anyway. And if we set the password, then unlock, we
simply skip unlocking, because setting the password implicitly unlocks
as well. The former case already worked, this commit adds support for
the latter.
(cherry picked from commit 3af3ec92c472e69da1096d8bfcb15fb08bf2bc9f)