Make serial tests into "barriers" and enable "serial" in the filename.
"Barriers" mean that a serial test will now execute in between other
tests, but will make sure that all tests before it have finished, and
all tests after it have not started yet, regardless of whether the
tests are parallel, serial or timed. This was changed in order to
allow serial tests to be used as initializers/finalizers. You can
use a serial test to do setup before a pile of other tests without
having to make all of them serial. Lexicographical sorting order
determines the order.

The old way was to have all serial tests executed at the end.