cf-monitord: Add support for temperature data on newer Linux hosts
Current distributions no longer use the ACPI interface in /proc.
With the old code, they need an installation of lmsensors.

The lmsensor interface is not always available, unreliable and
expensive because an external program is called. It is also broken
because at least on Debian systems it won't return the CPU temperature.

Add support for the new virtual thermal device in /sys.

As data from lmsensors is not always compatible with data from the
virtual thermal device, the new code might need fixes in the
cfengine configuration.

Signed-off-by: Stefan Weil <sw@weilnetz.de>