Enable support for Busybox tools.
When running on Linux, CFEngine core assumes that the GNU utilities are
used.  Howver on embedded devices we commonly (always?) use busybox.
Thus 'out-of-the-box' certain functions do not work.  For example, it's
not possible to bootstrap a device because busybox /bin/ps accepts
different arguments.  The patch requires __BUSYBOX__ to be defined in
the build environment for busybox targets - hopefully there is a better
way where it can be detected automatically at runtime.