Deal with special soft_fail/skip situation in user promise tests.
Because we are reusing the same test twice in user promises (once as
a 10_*.cf file, later as a 20_*.cf symlink to the same file), we might
sometimes both skip and soft_fail the same test simultaneously, and
the test framework doesn't like this. Since this is likely to be quite
rare, we fix it in the relevant tests with some extra conditions.
(cherry picked from commit 218158b45a0978055ea01a222d461efb4cf6a224)