Redmine#6946: Fix detection of stopped process.
While `pr_flags` always has correctly the flag PR_STOPPED set, it is not
always the case that `pr_why` is set to PR_SIGNALLED. In the
process_test it happened that `pr_why` was PR_JOBCONTROL. So we are
accepting all values in `pr_why` now, the process is stopped anyway so
it shouldn't matter. For the reference, the manual is not that
clear. According to Solaris man 4 proc:

 PR_SIGNALLED

    indicates that the lwp stopped on receipt of a signal.

 PR_JOBCONTROL

    indicates that the lwp stopped due to the default action of a job
    control stop signal.