Use PROCESS_STATE_ZOMBIE to avoid unnecessary re-signalling.
Previously if we called GracefulTerminate() for one of our children, we
would keep sending hundreds of signals for one second. Reason is that
nobody is reaping the child, and the inner loop of
ProcessWaitUntilExited() keeps signalling the zombie.

With previous commits code was added, so now we know when a process is a
zombie. So we're using that code to avoid unnecessary signalling.