Redmine #6401, #6402: Introduce proper systemd support in core.
Change systemd definition to include services in two groups: web and
non-web. The reason for this is so that web services can be launched
after installation without CFEngine being bootstrapped yet. The main
cfengine3 service will launch the web service if it is available.

Also, move the init script and systemd service files to core, so they
can be installed with "make install". This is easier than letting the
package scripts do it. By default it is off so users don't overwrite
their init script files unintentionally when developing (since
--prefix is not honored for those files.