Redmine #7045: Restore RHEL7 hub functionality after 8668c38b69.
Move cfengine3-web.service to core. We don't need it there, but
systemd makes it impossible to make a conditional dependency on a
service, so we need to keep the file in core and put the conditional
inside the actual service instead.

Also add "After=" directive so that we know the database is up by the
time cf-consumer wants to access it.