Make backgrounding in tests easier by introducing no_fds command.
The problem is that we don't know which file descriptors have been
opened behind our back, and most will be inherited by the new process.
The original process or its parent may then hang because some file
descriptor is still open, even though we didn't intend that. The
FD_CLOEXEC mechanism is not very helpful for descriptors we don't own,
so use the closefrom call to close all the descriptors we don't need
in the new process instead.