Revert "Redmine #7042: In 3.6 only use systemd on RedHat."
This reverts commit 009bc4862f4262015df7102426e9e144dec2b16c.

In fact it works better if the code path is not gated to RedHat since
it will then call the init script and still keep all the daemons
within the correct systemd cgroup. The alternative, to launch the
daemons ourselves, would expose the old bug of ending up with the
daemons in the wrong cgroup.

Note that the Enterprise hub will not work with either method, because
it needs the support files for the web services, which are ATM not
present in 3.6 packages other than for RHEL7.