A key must now be loaded at the beginning of the agent run.
For example, if we try to initialise a TLS connection but a key is not
loaded, we will fail. Previously, it was possible to create a key
*while* the system was converging, so the first iteration it would fail,
the second iteration it would succeed.

The reason for changing this is that proper TLS initialisation involves
much more than just reading the keys (for example we need to set ciphers
and TLS version to be used), thus it can only happen during program
initialisation where that information (usually configured in body common
control) is easily accessible.

Most obvious side-effect in practice: "cf-agent --bootstrap".

If you have no key generated while bootstrapping, the bootstrap
procedure can't generate one for you, but it should inform you to
generate on by running cf-key, and then retry to bootstrap.