Rephrased message when abandoning collect call attempt.
Mentioning the parameter, collect_window, that controls whether to do
this gives the admin some chance of tweaking it, if that might help.
From the user's perspective, we're abandoning the call; the fact that
we're closing the socket is an implementation detail; so change
wording to say "abandon" instead of "close".