Add missing call to CollectCallMarkProcessed();
When CollectCallHasPending()/CollectCallGetPending() have been used to
take over ownership of the pending CC socket, their caller has a duty
to CollectCallMarkProcessed() when done with it.  That was all handled
suitably when actually using the socket; but was neglected when
discarding it due to an over-long queue.  This would leave a client
permanently failing to attempt collect calls.  Added the missing call.