Avoid signal pipe leak if we fail to set it non-blocking.
We carefully ensure the signal pipe shall be closed when we exit;
however, as we're setting it up, we exit() if we can't mark it
non-blocking; so we should set the atexit() handler *before*
attempting to so mark it, not after.  The need to close it is relevant
from the moment we open it, not from after we've got it nicely set up.

(Noticed in the course of preparing to test the PassOpenFile API, as
the test code needed a pipe pair, so I looked here to see how that's
done.)