Avoid duplication of common code and fix an inaccurate message.
The bulk of how we handle failure to set a socket non-blocking is the
same between MingGW and the free world, so use only one copy of the
code to do it.  The error message needed to include the name of the
function called, but a local macro easily takes care of that.

The error message in this code claimed the problem was with opening
the socket pair, where actually it was with setting it non-blocking.
Tell the truth.  Also split the message across lines, as it was long.