Fix old packages promise generic package_method handling.
Instead of adding dummy constraint while parsing policy whole logic
is moved to checking promise attributes. This needs a
CopyBodyConstraintsToPromise() function to copy constraints form
given body to specific promise.

For the new package promise we need to have the name of the package_module
body. This was removed and replaced with binary value in current
implementation. In order to have the knowledge of the body name one extra
constraint is added to each promise containing body constraint which is
[body_name]_name. This can be used to figure out the name of specific body
constraint in promise.
Assuming we have 'package_module => apt_get' 'package_module_name' will
refer to 'apt_get'.
constraint in policy