redmine7219: clarify error message and condition on caring.
No column has overflowed anything.  We *are* on the point of
overflowing the table, if there's more columns to come - except that
we side-step and don't make that mistake.  In any case, filling the
last column of the table isn't an error; the only problem would come
if there were another column to come, for which we lack space in the
table.  So stop to see if there actually *is* another column to come,
before Log()ging any error.  Also, testing > limit - 1 is just an ugly
way to test >= limit !