Include the template files for extensions_template.[ch] in the source tarball
To achieve this the files were renamed out of their .[ch] extension,
because make was trying to compile them after listing them as a
dependency to libpromises. By giving the extension .c.pre, it's clear
that they are not C files, but files that need pre-processing to become
valid C files.

Also removed their mention in the root Makefile, it doesn't seem to be
necessary anymore.