Fix SIGTERM waking up the loop instead of ending it.
After all the sleeping we do, we need to make sure that we check if we
should terminate, otherwise the signal will simply spin the loop,
which triggers cf-agent, which spawns a new cf-execd, and we are back
where we started. At least by checking the flag, this probability is
drastically reduced.

Note that this does not fully fix the problem, since the signal
arrival is inherently a race condition, and this cannot be avoided. So
additional changes to the service scripts are still needed in order to
make sure that all daemons are stopped.

Also make sure that we do not enter yet another sleep after the first
has been interrupted. Again, this is a race condition, but it should
still improve things.

Redmine #7193 and #6659.