Add version and architecture to be send with get-package-info.
In the new implementation of package promise only package name
was sent to package module while sending 'get-package-info'. In case
of multiple package versions in the same package file or in repo (not the
case in dpkg and rpm) we might need to send package version and
architecture to indicate which versions we are interested in.

If 'get-package-info' returns version and/or architecture for repo type
package those are ignored and verbose message is printed.