Redmine #7149: Make sure file desc. inheritance is honored on Windows.
It's quite important because Windows does not allow you to do anything
to a file that is open for writing, and if the file descriptor is
inherited by a process that is long lived, you'll never be able to
delete it.

This particular case was related to cf-execd output logs, that could
not be rotated because of cf-agent inheriting a file descriptor from
cf-execd, and cf-agent then launching an unrelated and long lived
process.

Also remove our null definition of the FD_CLOEXEC flag. Since we now
have the wrapper, we don't want people to use it by mistake.