Redmine #7193: Rewrite init script in order to be much more robust.
This will help both when manually running the init script, and
especially when performing upgrades, which will also run the init
script to stop the processes. Hopefully processes that are left over
while performing the upgrade should be a thing of the past.

Also note the new POSIX shell detection block, the old one (found
throughout the CFEngine repositories), as far as I can tell, does not
work. We should probably replace other instances of this as well, but
I will do that separately.

And at last, the script finally gets a unit test. It's not 100%
realistic, since we are killing a test process, and not a system
process, but should still go a long way in proving that the script
works on different platforms.

Changelog: Fixed a bug which could cause daemons to not to be killed
correctly when upgrading or manually running "service cfengine3 stop".