Switch to more efficient rewriting of Rval variable references.
Instead of looping per possible key, only loop for each reference to a
variable we find. The number of iterations is guaranteed to be less
than or equal to the previous approach, since that would scan the
string twice for each possible key, now we scan it once regardless of
available keys.