Fix passopenfile_test that was failing on HP-UX.
The registered atexit() call was being called for each test, at the
exit() of the child process, and was removing the Unix domain socket at
/tmp. This happened on every child, no matter if it had setup a socket
or not, because DIALEDUP was being turned on around the beginning of the
parent process and never turned off, despite the file being deleted by
the child.

The solution is to clean up the Unix domain socket file after every
test, both for child (via atexit()) and parent (manually) processes. The
cleanup is accompanied by setting DIALEDUP=false, thus the next spawned
child has correct state and won't try to remove a file that hasn't
created itself.