Introduce new cf-runagent acceptance tests framework
It mostly consists of special commands and policy to check that
cf-serverd invokes the "cfruncommand" correctly. The "cfruncommand" is
invoking a special script "write_args.sh" that outputs its arguments to
a file. At the end of the acceptance tests, that file is compared to the
expected "cfruncommand" arguments, and the test fails or passes.