Add: Acceptance tests for abortbundle classes
Test that the agent aborts the current bundle when an exact abortclass
is matched and that the agent continues on to the next bundle.

Test that the agent aborts the current bundle when a regular expression
is matched and that the agent continues on to the next bundle.

Test that the agent aborts if a classexpression is matched and that the
agent continutes on to the next bundle.

Test that the agent does *NOT* abort if a classexpression does not
match.

Ref: https://dev.cfengine.com/issues/3611
(cherry picked from commit 5615b4cb8d7071c875556a4fc28f6cc52ef549d8)