Redmine #7606 Improve TTY detection
Right now if one of the stdin, stdout or stderr refers to a terminal
agent is considered to run as tty_interactive.

GenericAgentConfigNewDefault() is rewritten to take additional parameter
which is detected TTY state. This is enabling unit testing.

Policy loading is now unified for agent, server and monitor daemon. New
SelectAndLoadPolicy() function is added to check and load policy.

Tests for checking if failsafe.cf is correctly created depending if
we are running in tty_interactive mode added.

Changelog: TTY detection should be more reliable. (Redmine #7606)