Redmine #7191: Allow multiple -D and -N flags to binaries
Currently, if multiple -D or -N flags are specified on the command-line, this
causes all but the last to be ignored. These commits fixes this such that
multiple flags can be used, adding more classes to be raised and lowered.

Changelog: cf-agent, cf-execd, cf-promises, cf-runagent and cf-serverd honor multiple -D, -N and -s arguments