Redmine #7460: Unable to edit the same value in multiple regions.
The issue was that even though we've been editing value in selected
region the whole file was checked to see if value was already edited.
This was broken by commit f8c7dca55ed8d99212564ea130666de4bd5e0931 when
check if line exists in region was added.

As a bonus some optimizations are added. Instead of iterating
through whole file we can now iterate only over selected region as
any operations will be done inside the region (no need to check whole
file).

Changelog: It is possible to edit the same value in multiple regions
of one file. (Redmine #7460)