Add extensive debugging output to cf_lock operations.
The motivation is to try to track down really hard-to-find bugs where
CFEngine hangs on a lock.

The reason for not putting the code directly in the DB operations
themselves is that we want the cleartext keys for the locks, and they
are only available higher up in the callstack.
(cherry picked from commit 318d871fefe009030d2f7aee0cd539f382405430)