Created new param, channel_deployment_dir
to be set in the params.sh file for the GIT_POLICY_CHANNELS vcs type.
Updated readme.org with this info also.

Add mkdir command for channel_deployment_dir so we don't fail easily.

Add -x flag to git clean command to prevent any issue with a .gitignore file
in the staging dir causing the deployed code to differ from the intended
code from the git repo.