In case of networking error, assume checksum is wrong
When a checksum is done to check is it is necessary to update the file,
the copy will stop anyway because of the networking error.
When it is used to check a file after copy (before moving .cfnew)
the agent should not consider the content it has received as correct
(espacially in a networking/server error context).

Changelog: Title