Redmine #7808: Do not drop privileges when changes permissions on non fifo
The reason is that dropping privileges has some unexpected
consequences: CFEngine operates recursively on a file tree in a way
that may leave the directory being operated on last (e.g. files inside
the directory are treated before the directory itself). This means
that if we drop privileges, we may not always have the permissions to
enter the directory. This is impossible to fix for FIFOs (other than
changing the order that CFEngine does things, but that is on a
different code level and has other implications), because FIFOs will
block if we try to open a file descriptor to them, but we can fix it
for regular files by switching back to the old method of opening the
file descriptor and doing fchmod on that.

An additional level of complexity in this case is that the path may
end in a symlink *pointing* to a FIFO, hence the code that opens the
true parent directory of the leaf.

Also rename the CHMOD_SETEUID_WORKS flag to CHMOD_SETEUID_HIGHBITS_WORK
to make it clearer that it is not seteuid support which is missing,
but support for seteuid together with high bit flags in permission
modes.

Changelog: Fix a regression which would sometimes cause "Permission
denied" errors on files inside directories with very restricted
permissions.

Based on a patch by Alexis Mousset <alexis.mousset@normation.com>.